requirements.txt
google-generativeai==0.7.2
pillow==10.4.0
streamlit==1.37.1
streamlit-option-menu==0.3.13
# RUN - pip install -r requirements.txt
